Fix CI red on main: ruff lint and roadmap heading
Three issues found while auditing the C1-C10 / D1 work that landed directly on main without going through PR review:

- backtester.py: int(round(...)) is redundant (round() of a float already returns int)  RUF046 lint error in two rolling-window helpers

- backtester.py: rolling_lump_sum_annualized_returns signature exceeded the 120-char line and never went through ruff format

- roadmap-2026-q3.md: stray '---# ...' line at the top from a botched heading replacement, breaks markdown rendering

All three are pure cleanup with no behavior change. CI on main was failing since 0150a5a; this gets it back to green so future PRs have a reliable gate to merge against.

+++ backend/app/backtester.py
@@ -199,7 +199,7 @@ def rolling_annualized_returns(events: list[ContributionEvent], window_years: in
         return []
 
     ordered = sorted(events, key=lambda event: event.date)
-    window_days = int(round(365.25 * window_years))
+    window_days = round(365.25 * window_years)
     minimum_periods = max(4, window_years * 8)
     first_date = pd.Timestamp(ordered[0].date)
     points: list[tuple[str, float]] = []
@@ -224,12 +224,14 @@ def rolling_annualized_returns(events: list[ContributionEvent], window_years: in
     return points
 
 
-def rolling_lump_sum_annualized_returns(fixed_events: list[ContributionEvent], window_years: int) -> list[tuple[str, float]]:
+def rolling_lump_sum_annualized_returns(
+    fixed_events: list[ContributionEvent], window_years: int
+) -> list[tuple[str, float]]:
     if len(fixed_events) < 3 or window_years <= 0:
         return []
 
     ordered = sorted(fixed_events, key=lambda event: event.date)
-    window_days = int(round(365.25 * window_years))
+    window_days = round(365.25 * window_years)
     minimum_periods = max(4, window_years * 8)
     first_date = pd.Timestamp(ordered[0].date)
     points: list[tuple[str, float]] = []
